+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - show comments by default in linear issue view, use --no-comments to disable
 - project list command to display projects in a table format
 - project view command to show detailed project information
+- team list command to display teams in a table format
 
 ## [0.6.4] - 2025-08-12